+++ src/ua/ithillel/lesson5/Main.java
@@ -1,4 +1,4 @@
-package ua.ithillel.lesson4;
+package ua.ithillel.lesson5;
 
 public class Main {
     public static void main(String[] args) {
