+++ src/ua/ithillel/lesson5/Student.java
@@ -1,4 +1,4 @@
-package ua.ithillel.lesson4;
+package ua.ithillel.lesson5;
 
 final public class Student extends Society implements Tasks {
