+++ src/ua/ithillel/lesson5/UniversityInfo.java
@@ -1,4 +1,4 @@
-package ua.ithillel.lesson4;
+package ua.ithillel.lesson5;
 
 import java.util.Arrays;
